Update VarDecl to store a potentially-compound name
This does not include any changes to parsing, so it’s still an error to declare a variable like `var foo(x:y:)`.

This is mainly a refactor to update various parts of the compiler to expect a `DeclName` as the name of a `VarDecl` rather than an `Identifier`.

Where the change necessary to support compound names was obvious/trivial, the change has been made inline. In places where more refactoring or further thought is required, an appropriate assertion was added in addtion to a “TODO(Compound variable names)”, and support will be added in a later PR.