[WASM] Fix lazy global initializer signature
Lazy global init is used with swift_once which passes a context
pointer parameter. If lazy global init doesn't take a context argument,
caller and callee signatures are mismatched and it causes runtime
exception on WebAssembly runtime. So we need to add dummy argument
to consume the context pointer.
See also: emitLazyGlobalInitializer