[NFC] [ClangImporter] Refactor canonical enum case
Extracts a helper type to encapsulate the table used to canonicalize cases, and simplifies the logic by largely merging the handling of unavailable cases with that of canonical cases.

NOTE: In theory, this could cause a behavior change if canonicalEnumConstants was ever used to look up the same case twice. (Previously, `unimported` would have been nullptr, which would cause it to go down the “case is an alias” path; now, it would effectively be the same as `canonical` used to be, which would not go down that path.) I don’t believe this ever actually happened, but I’m leaving a breadcrumb in this commit message in case a bug is ever bisected to this commit.

# Conflicts:
#	lib/ClangImporter/ImportDecl.cpp