[CodeCompletion] Migrate unresolved member completion to make use of typeCheckForCodeCompletion
This hooks up unresolved member completion (i.e. `.<here>`) to the
typeCheckForCodeCompletion API to generate completions from all solutions the
constraint solver produces (even those requiring fixes), rather than relying on
a single solution being applied to the AST (if any). This lets us produce
unresolved member completions even when the contextual type is ambiguous or
involves errors.

Whenever typeCheckExpression is called on an expression containing a code
completion expression and a CompletionCallback has been set, each solution
formed is passed to the callback so the type of the completion expression
can be extracted and used to lookup up the members to return.