Sema: Tweak preCheckExpression()'s use-before-def check for code completion
It's possible that the definition is in the original source buffer,
but the use is in a replaced range, in which case we would randomly
reject it based on the memory locations of the buffers.

I don't have a test case for this with parse-time lookup on;
with parse-time lookup off, it reproduces readily in all the
code completion tests in test/SourceKit/CodeCompletion.