[Structural] Add _Structural experimental module to stdlib.
This is a first commit of a series introducing automatic requirement
satisfaction, as described in the proposal "Automatic requirement satisfaction"
shared on the Swift forums. This work was originally prototyped at
google/swift-structural, with tests and benchmarks focusing on end-to-end case
studies.

This PR introduces an experimental `_Structural` standard library module
containing a single `Structural` protocol, used as the entry point to structural
generic programming. It does not yet include the complete API surface. Code
owners have given approval to upstream this code behind an
`--enable-experimental-structural` flag.

This change focus solely on adding the module with the corresponding conditional
compilation directives in the build and unit tests. Follow-up PRs will add
additional APIs and corresponding compiler support.

Co-authored-by: Dan Zheng <danielzheng@google.com>