[ownership] Teach the ownership verifier how to verify that guaranteed yielded values are only used within the coroutine's lifetime.
I think validating this was an oversight from the bringup of coroutines. I
discovered this while writing test cases for coroutine lifetime extension. I
realized it was possible to write a test case that should have triggered this
but was not.

I added some tests to make sure that we continue to flag this in the future.

rdar://69597888