Added a simple check for a period token after the identifier to inhibit the IsProbablyFuncDecl logic.
When the parser finds an identifier at type scope it tries to determine
if it is a function declaration without the `func` keyword. This
generates a cascate of diagnostics if the identifier is actually an
assignment to part of a value using the period (like a.b = 3). In this
case the parser should not try to treat is as a function and instead
just generate a simple expected declaration diagnostic.