Fetch associated types abstractly when fetching an associated conformance.
By default, emitTypeMetadataRef does a blocking request for complete
metadata, which is the right thing to do for most purposes in IRGen.
Unfortunately, it's actively dangerous in code that can be called
during metadata completion, like an associated conformance accessor,
because it can cause artificial dependency cycles that the runtime
isn't equipped to detect, much less solve.

This is a partial fix for rdar://69901318, which also exposes a bad
metadata access path that seems to be causing an artificial problem.