[Concurrency] Rename IsolationRestriction and use it more widely.
Rename IsolationRestriction to ActorIsolationRestriction. This type
governs the *use* of a potentially actor-isolated entity, which
includes the notion that (e.g.) an asychronous function can be used
from any context even when it is actor-isolated. Use this for @objc
checking and protocol conformance checking, in addition to its
original use for enforcing actor isolation in statements and
expressions.